add MMModemAccessTechnology and MMModem3gppRegistrationState enums
These aren't directly used in the interface code yet, but can be used by
consumers of the library to interpret the results from the
simple interface get_status() call.

MMModem3gppRegistrationState is also used in the
ModemManager1.Modem.Modem3gpp interface which is not yet
implemented.

MMModemAccessTechnology is used in ModemManager1.Modem
.AccessTechnologies which is only implemented for the async case yet.